Restructuring app by splitting shell
