+++ moola/shell/gsheets.py
@@ -1,68 +1,15 @@
 import json
-import os
 
-import click
 import gspread
 from gspread.exceptions import WorksheetNotFound
 from oauth2client.client import SignedJwtAssertionCredentials
 
-from .shell_monzo import _get_current_balance
-from .core import daily_balances_for_month
-from .models import Transaction
-from .utils import (
+from ..models import Transaction
+from ..utils import (
     current_day,
     current_month_number,
     current_year,
     get_worksheet_name)
-
-
-@click.command()
-@click.option(
-    '--environment',
-    default='production',
-    prompt='Spreadsheet to use',
-    type=click.Choice(['development', 'test', 'production']))
-@click.option(
-    '--year',
-    default=current_year(),
-    prompt='Year to use')
-@click.option(
-    '--month',
-    default=current_month_number(),
-    prompt='Number of month to use (1=Jan, 2=Feb, etc)',
-    type=click.IntRange(1, 12))
-@click.option('--start', default=2500.00, prompt='Start Balance')
-@click.option('--end', default=500.00, prompt='End Balance')
-def create_sheet_for_month(year, month, start, end, environment):
-    """
-    Prompt user for required values then create Google spreadsheet with amounts
-    for month
-    """
-    # TODO: validate year values here, could be any integer
-    if environment == 'production':
-        name = 'Money 2016'
-    elif environment == 'test':
-        name = 'Money test'
-    else:
-        name = 'Money dev'
-
-    spreadsheet = _get_google_spreadsheet(name)
-    transactions = _get_monthly_transactions(spreadsheet)
-
-    balances = daily_balances_for_month(
-        year,
-        month,
-        start,
-        end,
-        transactions)
-
-    worksheet = _write_balances_to_spreadsheet(
-        spreadsheet,
-        balances,
-        year,
-        month)
-    _print_spreadsheet_url(spreadsheet)
-    return worksheet
 
 
 def delete_worksheet(year, month):
@@ -78,23 +25,6 @@ def delete_worksheet(year, month):
         pass
 
 
-def get_monzo_balance():
-    """
-    Get balance from Monzo and write to Google Sheet
-    """
-    balance = _get_current_balance(
-        account_id=os.environ.get('MONZO_ACCOUNT_ID'),
-        access_token=os.environ.get('MONZO_ACCESS_TOKEN'))
-    print('balance in pence {}'.format(balance))
-
-    # TODO: toogles on this for dev, etc
-    spreadsheet = _get_google_spreadsheet('Money 2017')
-    _write_monzo_balances_to_spreadsheet(spreadsheet, balance)
-    _print_spreadsheet_url(spreadsheet)
-    # TODO: write to google_sheet
-
-
-# TOOD: move functions to shell/gsheets.py
 def _print_spreadsheet_url(spreadsheet):
     url = 'https://docs.google.com/spreadsheets/d/{0}/edit'.format(
         spreadsheet.id)
