+++ requirements.txt
@@ -1,4 +1,5 @@
 # Testing
+freezegun==0.3.9
 mock==1.3.0
 pytest==2.8.7
 responses==0.7.0
